refactored to a single fragment in packet mode (removed TabLayout). Added UI in packet mode fragment: init receive/transmit buttons, modulation spinner, manchester encoding slider, data rate text input, payload text input, buttons to send/receive
