--- app/src/main/java/com/example/emwaver10/ui/packetmode/ViewPagerAdapter.java
@@ -1,40 +1,0 @@
-package com.example.emwaver10.ui.packetmode;
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
-import androidx.viewpager2.adapter.FragmentStateAdapter;
-
-import java.util.List;
-import java.util.ArrayList;
-
-public class ViewPagerAdapter extends FragmentStateAdapter {
-
-    private final List<Fragment> fragments = new ArrayList<>();
-
-    private final List<String> fragmentTitles = new ArrayList<>();
-
-    public ViewPagerAdapter(@NonNull FragmentActivity fragmentActivity) {
-        super(fragmentActivity);
-    }
-
-    public String getTitle(int position){
-        return fragmentTitles.get(position);
-    }
-
-    public void addFragment(Fragment fragment, String title){
-        fragments.add(fragment);
-        fragmentTitles.add(title);
-    }
-
-    @NonNull
-    @Override
-    public Fragment createFragment(int position) {
-        return fragments.get(position);
-    }
-
-    @Override
-    public int getItemCount() {
-        return fragments.size();
-    }
-}
