--- app/src/main/res/layout/dropdown_item.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-    <TextView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/textViewModulations"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="15dp"
-        android:text=""
-        android:textColor="@color/black"
-        android:textSize="18sp"
-        android:textStyle="bold" />
