+++ app/src/main/res/layout/fragment_packet_mode.xml
@@ -7,35 +7,101 @@
     tools:context=".ui.packetmode.PacketModeFragment">
 
 
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tabLayout"
-        android:layout_width="409dp"
-        android:layout_height="wrap_content"
+    <Button
+        android:id="@+id/sendTesla"
+        android:layout_width="114dp"
+        android:layout_height="55dp"
+        android:text="send tesla"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.1" />
 
-        <com.google.android.material.tabs.TabItem
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/receive_tab" />
+    <androidx.appcompat.widget.SwitchCompat
+        android:id="@+id/manchesterSwitch"
+        android:layout_width="130dp"
+        android:layout_height="48dp"
+        android:text="Manchester encoding"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/initReceiveButton" />
 
-        <com.google.android.material.tabs.TabItem
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/transmit_tab" />
 
-    </com.google.android.material.tabs.TabLayout>
+    <AutoCompleteTextView
+        android:id="@+id/modulationSelector"
+        android:layout_width="100dp"
+        android:layout_height="48dp"
+        android:focusable="false"
+        android:focusableInTouchMode="false"
+        android:hint="Modulation"
+        app:layout_constraintEnd_toStartOf="@+id/datarateTextInput"
+        app:layout_constraintStart_toEndOf="@+id/manchesterSwitch"
+        app:layout_constraintTop_toBottomOf="@+id/sendTesla" />
 
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/viewPager2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent"
+
+    <EditText
+        android:id="@+id/datarateTextInput"
+        android:layout_width="161dp"
+        android:layout_height="48dp"
+        android:background="#EEE"
+        android:hint="Enter data rate..."
+        android:imeOptions="actionDone"
+        android:inputType="number"
+        android:padding="8dp"
+        android:textColor="#000"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintTop_toBottomOf="@+id/initTransmitButton" />
+
+    <Button
+        android:id="@+id/sendPayloadButton"
+        android:layout_width="114dp"
+        android:layout_height="55dp"
+        android:text="SEND PAYLOAD"
+        app:layout_constraintEnd_toStartOf="@+id/receivePayloadButton"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tabLayout" />
+        app:layout_constraintTop_toBottomOf="@+id/payloadDataTextInput" />
+
+    <Button
+        android:id="@+id/receivePayloadButton"
+        android:layout_width="114dp"
+        android:layout_height="55dp"
+        android:layout_marginEnd="60dp"
+        android:text="receive payload"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/payloadDataTextInput" />
+
+    <Button
+        android:id="@+id/initTransmitButton"
+        android:layout_width="114dp"
+        android:layout_height="55dp"
+        android:text="init transmit"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/sendTesla"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/initReceiveButton"
+        android:layout_width="114dp"
+        android:layout_height="55dp"
+        android:text="init receive"
+        app:layout_constraintEnd_toStartOf="@+id/sendTesla"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+    <EditText
+        android:id="@+id/payloadDataTextInput"
+        android:layout_width="411dp"
+        android:layout_height="100dp"
+        android:background="#EEE"
+        android:gravity="top|start"
+        android:hint="Hex data received/to transmit..."
+        android:inputType="textMultiLine"
+        android:padding="8dp"
+        android:scrollbars="vertical"
+        android:textColor="#000"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/modulationSelector" />
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
