--- app/src/main/res/layout/fragment_receive.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.packetmode.ReceiveFragment">
-
-    <TextView
-        android:id="@+id/textReceive"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias="0.042" />
-
-    <TableLayout
-        android:id="@+id/tableLayout"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintTop_toBottomOf="@+id/textReceive"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/receiveDataButton"
-        android:stretchColumns="*" />
-
-    <Button
-        android:id="@+id/receiveDataButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="receive"
-        app:layout_constraintTop_toBottomOf="@+id/tableLayout"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/receiveInitButton"
-        android:layout_width="155dp"
-        android:layout_height="88dp"
-        android:text="Button"
-        app:layout_constraintTop_toBottomOf="@+id/receiveDataButton"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias="0.1" />
-
-
-    <EditText
-        android:id="@+id/preambleTextInput"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:hint="Insert the num preamble bytes"
-        android:background="#EEE"
-        android:textColor="#000"
-        android:padding="8dp"
-        android:imeOptions="actionDone"
-        android:inputType="number"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
