--- app/src/main/res/layout/fragment_transmit.xml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.packetmode.TransmitFragment">
-
-    <Button
-        android:id="@+id/sendButtonTransmit"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="send tesla"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias="0.1"/>
-
-    <Switch
-        android:id="@+id/manchesterSwitch"
-        android:layout_width="92dp"
-        android:layout_height="68dp"
-        android:textOff="No Manchester"
-        android:textOn="Manchester"
-        app:layout_constraintTop_toBottomOf="@+id/sendButtonTransmit"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias="0.05" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:id="@+id/bitrateTextInputLayout"
-        android:layout_width="0dp"
-        android:layout_height="50dp"
-        app:layout_constraintTop_toBottomOf="@+id/manchesterSwitch"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/setDataRateButton"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintWidth_default="wrap">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/bitrateTextInputEditText"
-            android:layout_width="146dp"
-            android:layout_height="72dp"
-            android:hint="Insert the bitrate"
-            android:inputType="number" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <Button
-        android:id="@+id/setDataRateButton"
-        android:layout_width="114dp"
-        android:layout_height="56dp"
-        android:text="set datarate"
-        app:layout_constraintTop_toBottomOf="@+id/manchesterSwitch"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/bitrateTextInputLayout"
-        app:layout_constraintHorizontal_bias="0.5" />
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.ExposedDropdownMenu"
-        android:id="@+id/modulationTextInputLayout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@+id/bitrateTextInputLayout"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:hint="Modulation">
-
-        <AutoCompleteTextView
-            android:id="@+id/autoCompleteTextView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:inputType="none"
-            android:textColorHint="@drawable/hint_selector"
-            android:text="Choose the modulation" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:id="@+id/payloadTextInputLayout"
-        android:layout_width="0dp"
-        android:layout_height="50dp"
-        app:layout_constraintTop_toBottomOf="@+id/modulationTextInputLayout"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/sendPayloadButton"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintWidth_default="wrap">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/payloadTextInputEditText"
-            android:layout_width="132dp"
-            android:layout_height="82dp"
-            android:hint="Insert the payload" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <Button
-        android:id="@+id/sendPayloadButton"
-        android:layout_width="114dp"
-        android:layout_height="55dp"
-        android:text="SEND PAYLOAD"
-        app:layout_constraintTop_toBottomOf="@+id/modulationTextInputLayout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/payloadTextInputLayout"
-        app:layout_constraintHorizontal_bias="0.5" />
-
-    <Button
-        android:id="@+id/initTransmitButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="init transmit"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias="0.9" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
